Changed tile component parent class
Due to the need of a box component for query checks with player mouse clicks, exit generators, etc and not wanting users to need an additional component, the wise thing to do is to simply make the TileComponent a child of UBoxComponent which is still derived from UActorComponent. This allows me to combine the box component and the tile functionality into one component.