Implemented simplex noise for tile "heights"
Using a noise generation library (FastNoiseLite), I have implemented new a couple new variables for each tile by generating a value using Simplex Noise. When a tile is created, using its position in the MapTiles array, the tiles noise value is found and can have various uses. In my game these noise values will be used to represent the height of a tile in the environment with tiles being split between 3+ "Height Levels" depending on their noise value. Units will either gain or lose accuracy depending on the height of the tile the unit is on, and the tile the target is on (with larger height differences giving more drastic values).