Created function library, teleport point class and interactable interface
The function library I have created houses a number of functions that will be used in various classes. This includes functions to round a position to the nearest multiple of tile size: fitting it nicely into the tile grid, functions to get a pointer to specific tiles given their location, and more. The teleportation point class can have a number of uses as it is quite generic, but its main use at the moment is to teleport the player's character between rooms which still needs to be implemented. The interact interface features functions that an Actor can implement for when the tile it is occupying is clicked, hovered, etc, so that we don't need a reference to the specific object to interact with it.