Pathfinding implemented, refactored tile highlight/colour adding code
Pathfinding has been nearly fully implemented. A path is now able to be found between 2 points and currently I am highlighting the tiles in the path with a unique colour. In the future, I will create a "Move" function in the PathfindingInterface which actors can implement to provide specific movement functionality. Also made some changes to the code that adds a highlight colour to tiles when interacted with, and fixed the issue of unhovering removing the highlight of already highlighted tiles. I fixed this by adding an integer "HighlightCounter" which counts how many times there has been an attempt to highlight the tile. Interactions to disable the highlight subtract from this variable and only once it reaches 0 can the highlight be disabled. There is an issue though of this variable having incorrect values that causes it to not disable tile highlight when it should which I need to fix.