Fixed small errors
Fixed the error concerning the SubtractTileColour() function which was caused by a tiny spelling mistake. The function declaration in the IsTile interface and all calls to the function were spelled correctly, however the function implementation in the TileNode class was spelled incorrectly (Sutbract vs Subtract) so no definition was found when the correct function was called. Also fixed an error concerning the A* pathfinding algorithm. This error I think was caused by the tile distance values produced by the GetDistance() function. To fix it, I simply increased the tile distance constants from 1.4 to 14 and 1 to 10.