Reimported missing materials, added to character pathfinding
I migrated the missing materials I need for the tiles which gave them the ability to be highlighted. I also worked on the character pathfinding component and its interaction with the action system. Current issue however is that after the player character has been moved, attempting to use the Move action again does not work.