Removed and changed unneeded code and fixed OnHover bug
Going through some of the code I have written, I realised some of it was useless so it has either been removed or commented out. During the code changes, in an attempt to have action functionality completely abstracted to their own class functions (and not the controller), due to how passing variables to functions work, I broke the OnHover functionality. I simply moved the value assigning of the CurrentlyHoveredTileComponent back to the controller. I will have a look back at this in the future once things are properly working to see how I can move it back to the action classes.