Progressed fixing tile highlight bug
An error that caused tiles not to be unhighlighted when they should has nearly been fixed by simply decrementing the tile HighlightCounter regardless of if there is still a colour in the colour history. Previously I was only decremnting the HighlightCounter when there was no colours left so the highlight should be disabled (DisableHighlight() function) however because I still need to decrement the counter when unhovering for example, I should decrement it every time SubtractColour is called.