Fixed issue of UI not "blocking" hover linetrace
While the player's mouse is over certain UI, I want the player not be able to hover  over tiles (mouse essentially ignoring UI) so I set up a couple delegates and functions in the Action base class (Disable/EnableHover) and StrategyController (Disable/EnableHovering). The combination of these functions means that when the unreal UI's inbuilt FOnHover/FUnHoverEvent function's are bound and called, they also call the relevant functions in the previously mentioned classes to enable and disable the bCanHover bool.