Working on fixing teleporter
Due to how unit movement works, when the unit is teleported they will continue to move back towards the selected tile (tile that is occupied by the teleporter. The fix will simply to call  the units StopMovement function and to occupy the current tile after teleportation but I need to move the base PathfindingComponent and the related IHasPathfinding interface to the map generation plugin because that is where the teleporter is. Plus I think the base component and interface should've been here from the start.