Created action data struct, and moved general functionality from MoveActionButton to base ActionButton class
Title. Next I need to create a way to signal whose turn it is, either by just some text or an object (faded light pillar?) that is attached to the actor whose turn it currently is. Also need to find a way to set CurrentAction from the turn actors selected button.